separate digitize restore logic
